vote feature refactored to a reducer
