+++ src/context/CommentsContext.js
@@ -1,15 +1,15 @@
 import React from 'react';
 
-export const CommentsContext = React.createContext({comments: [{
-  postId: 4,
-  created: new Date(new Date() - (5 * 3600 * 1000)),
-  author: 'Josh',
-  body: 'We need to hook this up to make it work realtime, nest responses like quoting on whatsapp, and the ability to turn a comment into a vote or todo item'
-}]});
+export const CommentsContext = React.createContext();
 
 
 class CommentsProvider extends React.Component {
-  state = {};
+  state = {comments: [{
+    postId: 4,
+    created: new Date(new Date() - (5 * 3600 * 1000)),
+    author: 'Josh',
+    body: 'We need to hook this up to make it work realtime, nest responses like quoting on whatsapp, and the ability to turn a comment into a vote or todo item'
+  }]};
 
   addComment = body => {
     this.setState({ comments: this.state.comments.concat({
