+++ src/context/VoteContext.js
@@ -1,19 +1,18 @@
 import React from 'react';
 
-export const VoteContext = React.createContext({polls: {
-  1: {
-    title: 'New Post',
-    id: 1,
-    createdBy: 'Josh',
-    votes: 5,
-    deadline: '3 days',
-    fields: ['one', 'two']
-  }
-},
-createPoll: () => {}});
+export const VoteContext = React.createContext();
 
 class VoteProvider extends React.Component {
-  state = {};
+  state = { polls: {
+    1: {
+      title: 'New Post',
+      id: 1,
+      createdBy: 'Josh',
+      votes: 5,
+      deadline: '3 days',
+      fields: ['one', 'two']
+    }
+  }}
 
   createPoll = newPoll => {
     const newPolls = { ...this.state.polls };
